マップの移植完了

+++ MyGame/DirectXResourse/Line/Grid/Grid.h
@@ -15,7 +15,7 @@ private:
 	Vector3 m_axis_y;//y軸
 
 	Vector3 m_origin;//原点
-	int m_wonRowNam;//マスの合計
+	float m_wonRowNam;//マスの合計
 
 	Line* m_line;
